Add support for "go to source" from all baseline image types
