+++ Plugins/IntelliJ/src/main/kotlin/dev/testify/TestFlavor.kt
@@ -1,12 +1,17 @@
 package dev.testify
 
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiMethod
 import dev.testify.extensions.PAPARAZZI_ANNOTATION
 import dev.testify.extensions.PREVIEW_ANNOTATION
 import dev.testify.extensions.SCREENSHOT_INSTRUMENTATION
 import dev.testify.extensions.SCREENSHOT_INSTRUMENTATION_LEGACY
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtNamedFunction
+
+typealias FindSourceMethod = (imageFile: VirtualFile, project: Project) -> PsiMethod?
 
 data class GradleCommand(
     val argumentFlag: String,
@@ -22,6 +27,7 @@ enum class TestFlavor(
     val methodInvocationPath: (className: String?, methodName: String?) -> String,
     val testGradleCommands: GradleCommand,
     val recordGradleCommands: GradleCommand,
+    val findSourceMethod: FindSourceMethod
 ) {
     Testify(
         srcRoot = "androidTest",
@@ -38,7 +44,8 @@ enum class TestFlavor(
             argumentFlag = "-PtestClass=$1",
             classCommand = "screenshotRecord",
             methodCommand = "screenshotRecord"
-        )
+        ),
+        findSourceMethod = ::findTestifyMethod
     ),
 
     Paparazzi(
@@ -56,7 +63,8 @@ enum class TestFlavor(
             argumentFlag = "--rerun-tasks --tests '$1'",
             classCommand = "recordPaparazziDebug",
             methodCommand = "recordPaparazziDebug"
-        )
+        ),
+        findSourceMethod = ::findPaparazziMethod
     ),
 
     Preview(
@@ -74,7 +82,8 @@ enum class TestFlavor(
             argumentFlag = "--updateFilter '$1'",
             classCommand = "updateDebugScreenshotTest", // TODO: Need to parameterize for build variant
             methodCommand = "updateDebugScreenshotTest"
-        )
+        ),
+        findSourceMethod = ::findPreviewMethod
     )
 }
